Completed

+++ balance/api/views.py
@@ -1,18 +1,25 @@
-from users.models import Users
-from .serializers import RevenueSerializer, ReserveSerializer, OrderForUserSerializer, UpBalanceUserSerializer, ServiceSerializer, ServiceAddOrderSerializer, CurrentUserSerializer
-from rest_framework import generics
-from rest_framework import viewsets
-from rest_framework.response import Response
-from reserve.models import Service, Order, Reserve, Revenue
-from djoser.views import UserViewSet
 from django.shortcuts import get_object_or_404
+from djoser.views import UserViewSet
+from rest_framework import generics, viewsets
 from rest_framework.decorators import action, permission_classes
-from rest_framework.status import HTTP_400_BAD_REQUEST, HTTP_201_CREATED, HTTP_204_NO_CONTENT
-from .permissions import IsOwner
 from rest_framework.permissions import IsAdminUser
+from rest_framework.response import Response
+from rest_framework.status import (HTTP_201_CREATED, HTTP_204_NO_CONTENT,
+                                   HTTP_400_BAD_REQUEST)
+
+from reserve.models import Order, Reserve, Revenue, Service
+from users.models import Users
+
+from .permissions import IsOwner
+from .serializers import (CurrentUserSerializer, OrderForUserSerializer,
+                          ReserveSerializer, RevenueSerializer,
+                          ServiceAddOrderSerializer, ServiceSerializer,
+                          TransactionSerializer, UpBalanceUserSerializer)
 
 
-class CustomServiceViewSet(generics.ListCreateAPIView, viewsets.GenericViewSet,):
+class CustomServiceViewSet(generics.ListCreateAPIView,
+                           viewsets.GenericViewSet,):
+    """Кастомный вьюсет для просмотра, удаления и добавления услуг в заказ"""
     pass
 
 
@@ -33,7 +40,7 @@ class UserNewViewSet(UserViewSet):
             context={'request': request}
         )
         return Response(serializer.data)
-    
+
     """Резервирование средств на отдельном счете"""
     @permission_classes([IsOwner])
     @action(detail=False, methods=['post', 'get'])
@@ -49,11 +56,14 @@ class UserNewViewSet(UserViewSet):
                         total.balance -= total.total
                         total.save()
                         serializer = ReserveSerializer(
-                            data={'user': user.id, 'reserve_balance': total.total}
+                            data={'user': user.id,
+                                  'reserve_balance': total.total}
                         )
                         serializer.is_valid(raise_exception=True)
                         serializer.save()
-                        return Response(serializer.data, status=HTTP_201_CREATED)
+                        return Response(
+                            serializer.data,
+                            status=HTTP_201_CREATED)
                     else:
                         content = {'error': 'Данный заказ уже зарезервирован'}
                         return Response(content, status=HTTP_400_BAD_REQUEST)
@@ -63,33 +73,63 @@ class UserNewViewSet(UserViewSet):
             else:
                 content = {'error': 'Недостаточно средств на счете'}
                 return Response(content, status=HTTP_400_BAD_REQUEST)
-            
+
         """Просмотр зарезервированных средств"""
         if request.method == 'GET':
             serializer = ReserveSerializer(reserve, many=True)
             return Response(serializer.data)
-    
+
     """Признание выручки"""
-    @permission_classes([IsAdminUser])
     @action(detail=True, methods=['post'])
     def revenue(self, request, id):
-        price = get_object_or_404(Reserve, user_id=id)
-        orders = Order.objects.filter(owner_id=id)
-        balance = Reserve.objects.filter(user_id=id)
-        total = get_object_or_404(Users, id=id)
-        balance.delete()
-        serializer = RevenueSerializer(
-            data={'price': price.reserve_balance,
-                  'user': id,
-                  'service': [o.service.name for o in orders],
-                  },
-        )
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        total.total = 0
-        total.save()
-        orders.delete()
-        return Response(serializer.data, status=HTTP_201_CREATED)
+        if self.request.user.is_staff:
+            price = get_object_or_404(Reserve, user_id=id)
+            orders = Order.objects.filter(owner_id=id)
+            balance = Reserve.objects.filter(user_id=id)
+            total = get_object_or_404(Users, id=id)
+            balance.delete()
+            serializer = RevenueSerializer(
+                data={'price': price.reserve_balance,
+                      'user': id,
+                      'service': [o.service.name for o in orders]},
+            )
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            total.total = 0
+            total.save()
+            orders.delete()
+            return Response(serializer.data, status=HTTP_201_CREATED)
+        content = {'error': 'Признать выручку может только администратор'}
+        return Response(content, status=HTTP_400_BAD_REQUEST)
+
+    """Перевод средств другому пользователю"""
+    @action(detail=True, methods=['patch'])
+    def transaction(self, request, id):
+        user_from = get_object_or_404(Users, id=self.request.user.id)
+        user_to = get_object_or_404(Users, id=id)
+        serializer = TransactionSerializer(data=request.data)
+        print(serializer.initial_data['transaction'])
+        print(user_from.balance)
+        transaction = serializer.initial_data['transaction']
+        if user_from.id == user_to.id:
+            content = {'error': 'Невозможно перевести средства самому себе'}
+            return Response(content, status=HTTP_400_BAD_REQUEST)
+        else:
+            if user_from.balance >= transaction:
+                user_from.balance -= transaction
+                user_from.save()
+                user_to.balance += transaction
+                user_to.save()
+                serializer = TransactionSerializer(
+                    data={'message': 'Перевод успешно совершен',
+                          'balance': user_from.balance,
+                          },
+                )
+                serializer.is_valid(raise_exception=True)
+                return Response(serializer.data, status=HTTP_201_CREATED)
+            else:
+                content = {'error': 'На вашем счете недостаточно средств'}
+                return Response(content, status=HTTP_400_BAD_REQUEST)
 
 
 class UpBalanceUserUpdate(generics.UpdateAPIView):
@@ -102,7 +142,7 @@ class UpBalanceUserUpdate(generics.UpdateAPIView):
 
 
 class ServiceViewSet(CustomServiceViewSet):
-    """Просмотр всех услуг"""
+    """Просмотр, добавление и удаление услуг из заказа"""
 
     queryset = Service.objects.all()
     serializer_class = ServiceSerializer
@@ -139,12 +179,12 @@ class ServiceViewSet(CustomServiceViewSet):
                 obj.delete()
                 content = {'message': 'Услуга удалена из заказа'}
                 return Response(content, status=HTTP_204_NO_CONTENT)
-            content={'error': 'Данной услуги нет в заказе'}
+            content = {'error': 'Данной услуги нет в заказе'}
             return Response(content, status=HTTP_400_BAD_REQUEST)
 
 
 class RevenueListCreate(generics.ListAPIView):
-    """Признание выручки и добавление средств с отдельного счета"""
+    """Признание выручки и добавление средств в выручку с счета резерва"""
 
     queryset = Revenue.objects.all()
     serializer_class = RevenueSerializer
